Create FBProphet_Cali
Here we fit the model for the 7 different product categories (this time for Cali), note that the product names are still labeled with 'MEDELLIN', but this is only a label issue and it was for simplicity. As in the Medellin code, we export the yhat, yhatmean and yreal, so the dashboard does not read the model directly, but a table with the results.